requirements.txt
asgiref==3.10.0
Django==5.2.7
django-axes==8.0.0
et_xmlfile==2.0.0
gunicorn==23.0.0
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
